+++ Source/ImoutoRebirth.Navigator/ImoutoRebirth.Navigator/Settings.Designer.cs
@@ -12,7 +12,7 @@ namespace ImoutoRebirth.Navigator {
     
     
     [global::System.Runtime.CompilerServices.CompilerGeneratedAttribute()]
-    [global::System.CodeDom.Compiler.GeneratedCodeAttribute("Microsoft.VisualStudio.Editors.SettingsDesigner.SettingsSingleFileGenerator", "16.5.0.0")]
+    [global::System.CodeDom.Compiler.GeneratedCodeAttribute("Microsoft.VisualStudio.Editors.SettingsDesigner.SettingsSingleFileGenerator", "17.9.0.0")]
     internal sealed partial class Settings : global::System.Configuration.ApplicationSettingsBase {
         
         private static Settings defaultInstance = ((Settings)(global::System.Configuration.ApplicationSettingsBase.Synchronized(new Settings())));
@@ -97,13 +97,25 @@ namespace ImoutoRebirth.Navigator {
         
         [global::System.Configuration.UserScopedSettingAttribute()]
         [global::System.Diagnostics.DebuggerNonUserCodeAttribute()]
-        [global::System.Configuration.DefaultSettingValue("False")]
+        [global::System.Configuration.DefaultSettingValueAttribute("False")]
         public bool ShowSystemTags {
             get {
                 return ((bool)(this["ShowSystemTags"]));
             }
             set {
                 this["ShowSystemTags"] = value;
+            }
+        }
+        
+        [global::System.Configuration.UserScopedSettingAttribute()]
+        [global::System.Diagnostics.DebuggerNonUserCodeAttribute()]
+        [global::System.Configuration.DefaultSettingValueAttribute("False")]
+        public bool AutoShuffle {
+            get {
+                return ((bool)(this["AutoShuffle"]));
+            }
+            set {
+                this["AutoShuffle"] = value;
             }
         }
     }