+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Navigator
 * Move extensions to common project
+* Add settings for auto shuffle on every reload (default: false)
 
 ### Harpy
 * Add ability to download faved danbooru pictures through gelbooru